Trying to upgrade it as per Walter's comment:
Daksh, mechanically you have satisfied Sam's requests. But I am not convinced that the text you have written is very helpful to a newbie. For example, the discussion of the Frame should begin with What is the Frame? How do you access? Why do you want to access it? And much less important, How to change the settings.  You put settings upfront and only describe functionally the top left of the Frame.   I have similar comments for all of the other sections. Alas, more work needed.