GSOC 2017 project final code pull request
### Introduction
This is the pull request with the final code changes I made as part of
the GSOC 2017 project **Giving Sugar Labs Website a New Look**. The
changes made during the **GSOC** period can be viewed at
[https://github.com/sugarlabs/www-sugarlabs/issues/63](https://github.co
m/sugarlabs/www-sugarlabs/issues/63).
### Note
The work during **GSOC** was done in the repository
[https://github.com/geekrypter/sugarLabsWebsiteRedesign](https://github.
com/geekrypter/sugarLabsWebsiteRedesign). A live preview of the website
is hosted using **Github pages** and can be viewed at
[https://geekrypter.github.io/sugarLabsWebsiteRedesign/](https://geekryp
ter.github.io/sugarLabsWebsiteRedesign/). However the repository
[https://github.com/geekrypter/sugarLabsWebsiteRedesign](https://github.
com/geekrypter/sugarLabsWebsiteRedesign) is not a forked version of the
repository
[https://github.com/sugarlabs/www-sugarlabs](https://github.com/sugarlab
s/www-sugarlabs). So for the pull request, I created a forked
repository at
[https://github.com/geekrypter/www-sugarlabs](https://github.com/geekryp
ter/www-sugarlabs) and under the branch
**GSOC-2017-project-final-code**, I created this pull request.
### Commits
The link where you can see my commits is -
[https://github.com/geekrypter/sugarLabsWebsiteRedesign/commits/master](
https://github.com/geekrypter/sugarLabsWebsiteRedesign/commits/master).
You can check the timestamps to see the commits related to any
particular day.
### Documentation
For documentation, please check the **documentation.md** file.
### Future developement and improvements
1. The website has been targetted to work on as many browsers as
possible. However, it can be made more compatible for OLPC as there are
certain issues:
a. Please follow this [mail
thread](http://lists.sugarlabs.org/archive/sugar-devel/2017-August/05453
0.html) for the plugins issue and CPU usage issue.
b. This is the way the website is rendered on OLPC XO 1.5 running
latest Sugar as reported by Samson Goddy. -
[https://drive.google.com/file/d/0B0pU4Ahiyt7UeDdFYXdfczd4bE0/view?usp=s
haring](https://drive.google.com/file/d/0B0pU4Ahiyt7UeDdFYXdfczd4bE0/vie
w?usp=sharing)
whereas, this is the way the website is loaded on Fedora SOAS (This how
it should be loaded in "Browse" activity).
-[https://drive.google.com/open?id=0B5pN2u8B5LwcX1IteGl4czhoQjA](https:/
/drive.google.com/open?id=0B5pN2u8B5LwcX1IteGl4czhoQjA)
2. The website is crashing on Chromium 60.0.3112.89 (dev build) while
using Raspbian Stretch. This issue is quite odd and doesn't show proper
trace. It has to be further inspected and solved.