Better Search Bar
Chnages are to make search bar resposive by keeping it at a distance from the navigation bar, making provision for clicking anywhere to disable search when expanded and removing animation from mobile view.
Search bar is made responsive by measurin position of the search bar and navigation bar and adjusting distance between them. This inludes all new jquery statements excpet first three.
The first three are to ewduce the size search bar when clicked anywhere eles in the browser window except on the search bar.
There are changes in expander() and condenser() to remove animation in mobile view. In condenser(), width of input box is incresed to 75% from 70% to make it look better. The onFocus attribute is removed and is changed into jquery statement.