Remove search
The search box regularly does not render in the correct position; at
the moment, it renders just above and to the right of the Sugar Labs
logo.  The search function does nothing particularly useful, it
delegates to Google Search.  It would be best to leave searching to
the better user experience of search engines and browsers.

- remove the search box,

- remove the search feature,

- remove the supporting styles and documentation.

Signed-off-by: James Cameron <quozl@laptop.org>