Increased the Padding when navbar switches to banner-mobile
When the width of the page is less than 1215px, as stated in nav.html, the banner switches to banner-mobile, removing the Paypal/VISA/MasterCard image, which decreases the padding between the main page and navbar.

This fix increases the padding a little more(adjusting it dynamically) when the width is less than 1215px, giving more space for index.html

Fixes issue raised by pikurusa in the PR comment.